ai-service/requirements.txt
flask
flask-cors
requests
python-dotenv
torch
transformers
accelerate
bitsandbytes
